# Compile variable nodes into the Liquid::C::BlockBody VM code