--- a/ext/liquid_c/liquid.c
+++ b/ext/liquid_c/liquid.c
@@ -10,6 +10,10 @@
 #include "variable_lookup.h"
 #include "vm.h"
 
+ID id_evaluate;
+ID id_to_liquid;
+ID id_call;
+
 VALUE mLiquid, mLiquidC, cLiquidSyntaxError, cMemoryError, cLiquidVariable, cLiquidTemplate, cLiquidBlockBody;
 rb_encoding *utf8_encoding;
 int utf8_encoding_index;
@@ -21,6 +25,10 @@ __attribute__((noreturn)) void raise_non_utf8_encoding_error(VALUE string, const
 
 void Init_liquid_c(void)
 {
+    id_evaluate = rb_intern("evaluate");
+    id_to_liquid = rb_intern("to_liquid");
+    id_call = rb_intern("call");
+
     utf8_encoding = rb_utf8_encoding();
     utf8_encoding_index = rb_enc_to_index(utf8_encoding);