--- a/ext/liquid_c/liquid.h
+++ b/ext/liquid_c/liquid.h
@@ -5,6 +5,10 @@
 #include <ruby/encoding.h>
 #include <stdbool.h>
 
+extern ID id_evaluate;
+extern ID id_to_liquid;
+extern ID id_call;
+
 extern VALUE mLiquid, mLiquidC, cLiquidSyntaxError, cMemoryError, cLiquidVariable, cLiquidTemplate, cLiquidBlockBody;
 extern rb_encoding *utf8_encoding;
 extern int utf8_encoding_index;