--- a/ext/liquid_c/parser.h
+++ b/ext/liquid_c/parser.h
@@ -15,6 +15,7 @@ lexer_token_t parser_consume(parser_t *parser, unsigned char type);
 lexer_token_t parser_consume_any(parser_t *parser);
 
 VALUE parse_expression(parser_t *parser);
+bool will_parse_constant_expression_next(parser_t *parser);
 
 void init_liquid_parser(void);