--- a/ext/liquid_c/vm_assembler.c
+++ b/ext/liquid_c/vm_assembler.c
@@ -4,6 +4,8 @@ void vm_assembler_init(vm_assembler_t *code)
 {
     code->instructions = c_buffer_allocate(8);
     code->constants = c_buffer_allocate(8 * sizeof(VALUE));
+    code->max_stack_size = 0;
+    code->stack_size = 0;
 }
 
 void vm_assembler_free(vm_assembler_t *code)
@@ -22,13 +24,32 @@ void vm_assembler_gc_mark(vm_assembler_t *code)
     while (ip < end_ip) {
         switch (*ip++) {
             case OP_LEAVE:
+            case OP_POP_WRITE_VARIABLE:
                 break;
+
+            case OP_HASH_NEW:
+                ip++;
+                break;
+
+            case OP_RENDER_VARIABLE_RESCUE:
+                const_ptr++;
+                break;
+
             case OP_WRITE_RAW:
                 const_ptr += 2;
                 break;
+
             case OP_WRITE_NODE:
+            case OP_PUSH_CONST:
+            case OP_PUSH_EVAL_EXPR:
                 rb_gc_mark(*const_ptr++);
                 break;
+
+            case OP_FILTER:
+                ip++;
+                rb_gc_mark(*const_ptr++);
+                break;
+
             default:
                 rb_bug("invalid opcode: %u", ip[-1]);
         }