--- a/liquid-c.gemspec
+++ b/liquid-c.gemspec
@@ -18,6 +18,8 @@ Gem::Specification.new do |spec|
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ["lib"]
 
+  spec.required_ruby_version = ">= 2.5.0"
+
   spec.add_dependency 'liquid', '>= 3.0.0'
 
   spec.add_development_dependency 'bundler', ">= 1.5" # has bugfix for segfaulting deploys