~~Depends on https://github.com/Shopify/liquid-c/pull/58 and https://github.com/Shopify/liquid/pull/1294~~

This is a continuation of https://github.com/Shopify/liquid-c/pull/58 that compiles the strict parseable liquid variables directly in the block body VM code for performance.

For simplicity, variables and expressions in tags do not yet leverage this VM code and the variable VM code still relies on the existing expression evaluation code (e.g. `context_evaluate` of Liquid::VariableLookup and Liquid::RangeLookup).  These will be optimized in following PR(s).

All the commits except the (currently) last one (Compile variable nodes into the Liquid::C::BlockBody VM code) are refactors:
* ~~I've moved the VM instructions and constants out of block_body_t into a vm_assembler_t struct, since we will want to embed the VM code in a future Liquid::C::Variable.  In the future, I intend to wrap the assembler in a Liquid::C::Assembler as a safe API for compiling tags from ruby code~~ (edit: extracted to https://github.com/Shopify/liquid-c/pull/69)
* ~~I changed the c_buffer struct to store pointers for the end of the data and capacity, instead of storing the sizes and capacity directly.  This is because we are mostly getting and updating the data end pointer, which is especially true in order to use this for the VM stack where we reserve space for the block and don't have to check the capacity for each write/push.~~ (edit: extracted to https://github.com/Shopify/liquid-c/pull/70)
* ~~I've also added support to c_buffer to be initialized with size 0 (without an allocation) which is the one case where we can't just double the capacity until we have the requested extra capacity.~~ (edit: extracted to https://github.com/Shopify/liquid-c/pull/70)
* I introduced a Liquid::C::VM that is attached to the Liquid::Context through an hidden instance variable, so the stack can be shared with nested blocks
* Leveraging https://github.com/Shopify/liquid/pull/1294 to only run liquid integration tests to avoid having to be fully compatible with the deprecated Liquid::BlockBody#nodelist
* ~~I've added a liquid_vm_next_instruction function to reduce duplication for code iterating instructions by using it for both Liquid::C::BlockBody#remove_blank_strings and Liquid::C::BlockBody#nodelist.  This way I can further leverage it for variable error handling.~~ (edit: extracted to https://github.com/Shopify/liquid-c/pull/72)
* Drop support for ruby 2.4 which is no longer supported upstream

The VM stack is one of the significant additions from this PR.  I chose to re-use the stack for nested blocks, which means that it needs to be expandable, hence the use of c_buffer_t for its memory allocation.  We can't globally determine the maximum stack size, since this stack could even be re-used in a template partial, but this PR does calculate the maximum stack size needed for a block body so that we can reserve stack space at the start of block body render and not have to check for sufficient capacity throughout the block body render.

Another design choice I made for this PR is to do error handling for the whole block body render, rather than for each variable render.  This way we can reduce the state saving cost of `rb_rescue` from liquid code that doesn't encounter variable render errors.  In order to recover from the exception, we just restore the stack size to what it was at the start of the block body render and iterate the instructions to jump just past the variable write.

Since the variable lookup and expression evaluation is still happen as it did before, the most significant change for performance is the filter invocation, which can happen without allocating an arguments array by leveraging the VM stack calling the filter function directly using `rb_funcallv`.  The filter `invokable?` is still being done at runtime (we need to pass filters as a parse option to do it at parse time) but is still optimized by doing the check directly from C using a hash with symbol keys.

## Benchmark

Before this PR (on the https://github.com/Shopify/liquid-c/pull/58 branch)

```
              parse:    146.972  (± 2.0%) i/s -      1.470k in  10.007250s
             render:    162.266  (± 2.5%) i/s -      1.635k in  10.081595s
     parse & render:     70.513  (± 2.8%) i/s -    708.000  in  10.047458s
```

after

```
              parse:    163.206  (± 1.8%) i/s -      1.632k in  10.002316s
             render:    172.302  (± 4.1%) i/s -      1.734k in  10.080641s
     parse & render:     78.371  (± 1.3%) i/s -    784.000  in  10.006170s
```